refactor(test): getLogSpy를 src로 이동
__test__ 폴더 내의 모든 파일이 테스트 대상이 되기 때문에
테스트 유틸 함수를 src 폴더로 이동함.
