+++ __tests__/ApplicationTest.js
@@ -1,6 +1,6 @@
 import { MissionUtils } from "@woowacourse/mission-utils";
 import App from "../src/App.js";
-import getLogSpy from "./getLogSpy.js";
+import getLogSpy from "../src/getLogSpy.js";
 
 const mockQuestions = (inputs) => {
   MissionUtils.Console.readLineAsync = jest.fn();
