+++ src/getLogSpy.js
